requirements.txt
# HalluciNOT (LGP) — Core Dependencies
# Python 3.12.3

# ML / NLP
torch>=2.8.0
transformers>=4.57.0
sentence-transformers>=5.2.0
datasets>=4.5.0
accelerate>=1.11.0

# LLM APIs
google-generativeai>=0.8.0

# Symbolic / Math
sympy>=1.14.0
numpy>=2.2.0
scipy>=1.16.0
scikit-learn>=1.7.0

# Docker Sandbox
docker>=7.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tabulate>=0.9.0
